Add filter on data & Merge card in a tab
