+++ src/components/Main.tsx
@@ -1,11 +1,11 @@
 import React, { FC, useState } from "react";
 import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
-import InputLabel from "@material-ui/core/InputLabel";
-import MenuItem from "@material-ui/core/MenuItem";
-import FormControl from "@material-ui/core/FormControl";
-import Select from "@material-ui/core/Select";
-import TextField from "@material-ui/core/TextField";
-import Button from "@material-ui/core/Button";
+// import InputLabel from "@material-ui/core/InputLabel";
+// import MenuItem from "@material-ui/core/MenuItem";
+// import FormControl from "@material-ui/core/FormControl";
+// import Select from "@material-ui/core/Select";
+// import TextField from "@material-ui/core/TextField";
+// import Button from "@material-ui/core/Button";
 import SimpleTabs from "./SimpleTabs";
 
 const useStyles = makeStyles((theme: Theme) =>
@@ -97,8 +97,7 @@ const Main: FC = () => {
           입력
         </Button>
       </form> */}
-      {/* <SimpleTabs /> */}
-      <SimpleTabs />
+      <SimpleTabs title="title" category="category" />
     </div>
   );
 };
