+++ src/components/SimpleTabs/SimpleTabs.tsx
@@ -1,12 +1,15 @@
-import React from "react";
+import React, { FC, useContext } from "react";
 import { makeStyles, Theme } from "@material-ui/core/styles";
 import AppBar from "@material-ui/core/AppBar";
 import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
 import Typography from "@material-ui/core/Typography";
 import Box from "@material-ui/core/Box";
-import Cards from "../Cards";
-import { Card } from "@material-ui/core";
+import Grid from "@material-ui/core/Grid";
+import Card from "@material-ui/core/Card";
+import CardActionArea from "@material-ui/core/CardActionArea";
+import CardContent from "@material-ui/core/CardContent";
+import { FoodContext } from "../../context";
 
 interface TabPanelProps {
   children?: React.ReactNode;
@@ -37,21 +40,32 @@ const a11yProps = (index: any) => {
     "aria-controls": `simple-tabpanel-${index}`
   };
 };
+interface CardProps {
+  title: string;
+  category: string;
+}
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
     flexGrow: 1,
     backgroundColor: theme.palette.background.paper
+  },
+  card: {
+    margin: 0
   }
 }));
 
-const SimpleTabs = () => {
+const SimpleTabs: FC<CardProps> = ({ title, category }) => {
   const classes = useStyles();
   const [value, setValue] = React.useState(0);
-
+  const context = useContext(FoodContext);
   const handleChange = (event: React.ChangeEvent<{}>, newValue: number) => {
     setValue(newValue);
   };
+  const koreanFood = context.food.filter(item => item.category === "한식");
+  const chineseFood = context.food.filter(item => item.category === "중식");
+  const japaneseFood = context.food.filter(item => item.category === "일식");
+  const westernFood = context.food.filter(item => item.category === "양식");
 
   return (
     <div className={classes.root}>
@@ -68,16 +82,124 @@ const SimpleTabs = () => {
         </Tabs>
       </AppBar>
       <TabPanel value={value} index={0}>
-        <Cards title="title" category="한식"/>
+        <div className={classes.root}>
+          <Grid container spacing={2}>
+            {koreanFood.map((item, i) => (
+              <Grid item xs={4} key={i}>
+                <Card className={classes.card}>
+                  <CardActionArea>
+                    <CardContent>
+                      <Typography
+                        gutterBottom
+                        variant="subtitle2"
+                        component="p"
+                      >
+                        {item.category}
+                      </Typography>
+                      <Typography
+                        variant="h5"
+                        color="textSecondary"
+                        component="h2"
+                      >
+                        {item.title}
+                      </Typography>
+                    </CardContent>
+                  </CardActionArea>
+                </Card>
+              </Grid>
+            ))}
+          </Grid>
+        </div>
       </TabPanel>
       <TabPanel value={value} index={1}>
-        <Cards title="title" category="중식"/>
+        <div className={classes.root}>
+          <Grid container spacing={2}>
+            {chineseFood.map((item, i) => (
+              <Grid item xs={4} key={i}>
+                <Card className={classes.card}>
+                  <CardActionArea>
+                    <CardContent>
+                      <Typography
+                        gutterBottom
+                        variant="subtitle2"
+                        component="p"
+                      >
+                        {item.category}
+                      </Typography>
+                      <Typography
+                        variant="h5"
+                        color="textSecondary"
+                        component="h2"
+                      >
+                        {item.title}
+                      </Typography>
+                    </CardContent>
+                  </CardActionArea>
+                </Card>
+              </Grid>
+            ))}
+          </Grid>
+        </div>
       </TabPanel>
       <TabPanel value={value} index={2}>
-        <Cards title="title" category="일식"/>
+        <div className={classes.root}>
+          <Grid container spacing={2}>
+            {japaneseFood.map((item, i) => (
+              <Grid item xs={4} key={i}>
+                <Card className={classes.card}>
+                  <CardActionArea>
+                    <CardContent>
+                      <Typography
+                        gutterBottom
+                        variant="subtitle2"
+                        component="p"
+                      >
+                        {item.category}
+                      </Typography>
+                      <Typography
+                        variant="h5"
+                        color="textSecondary"
+                        component="h2"
+                      >
+                        {item.title}
+                      </Typography>
+                    </CardContent>
+                  </CardActionArea>
+                </Card>
+              </Grid>
+            ))}
+          </Grid>
+        </div>
       </TabPanel>
       <TabPanel value={value} index={3}>
-        <Cards title="title" category="양식"/>
+        <div className={classes.root}>
+          <Grid container spacing={2}>
+            {westernFood.map((item, i) => (
+              <Grid item xs={4} key={i}>
+                <Card className={classes.card}>
+                  <CardActionArea>
+                    <CardContent>
+                      <Typography
+                        gutterBottom
+                        variant="subtitle2"
+                        component="p"
+                      >
+                        {item.category}
+                      </Typography>
+                      <Typography
+                        variant="h5"
+                        color="textSecondary"
+                        component="h2"
+                      >
+                        {item.title}
+                      </Typography>
+                    </CardContent>
+                  </CardActionArea>
+                </Card>
+              </Grid>
+            ))}
+          </Grid>
+        </div>
       </TabPanel>
     </div>
   );
